Require tails of all files to associate any one chunk with a manifest
Resolved issue where it was possible to discern when any block was
associated with the same manifest file by comparing the hash blocks
within the manifest, which could be incriminating depending on the
content and how oppresive the governing regime of the user may be.

This was resolved by a feature that XOR's the hash blocks in the
manifest file with the last 32 bytes of all other encrypted pieces.
This means that it requires all pieces plus the manifest file to verify
the association of a block to a manifest file.